Made notes on making Bash Scripts
for a separate task, I needed to write a bash script which copies the context of a file called customCommands.txt to a main.tex file of a directory that needed to be an argument of the script. I thought i could make notes on how to do this on the notebooks of this directory as it may be useful to our project later.